지형, 배출구, 문, 레밍 객체 Actor 설정, 레밍 포지션 이동
스프라이트 애니메이션 재생 되는거 확인했으나,
픽셀 충돌과 레밍 이동 구현 코드 작성하다 재생이 멈춘 상태.
원인이 뭔지 알아보는 중.

+++ src/Game.h
@@ -4,8 +4,7 @@
 #include "GameScene.h"
 #include "GameState.h"
 #include "Texture.h"
-
-#include "SoundManager.h"
+#include "TimeManager.h"
 #include "MouseManager.h"
 #include "KeyboardManager.h"
 
@@ -30,7 +29,7 @@ public:
 	void init(HWND hwnd);
 	void destroy();
 
-	bool update(int deltaTime);
+	bool update(float deltaTime);
 	void render();
 
 	HWND getHwnd() { return _hwnd; }
